refactor: core 패키지 제거, 하위 패키지를 상위 디렉토리 레벨로 이동
