+++ app/src/main/java/com/boostcamp/and03/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.boostcamp.and03.ui.core.navigation.rememberMainNavigator
+import com.boostcamp.and03.ui.navigation.rememberMainNavigator
 import com.boostcamp.and03.ui.theme.And03Theme
 
 class MainActivity : ComponentActivity() {
