+++ app/src/main/java/com/boostcamp/and03/ui/navigation/MainBottomTab.kt
@@ -1,4 +1,4 @@
-package com.boostcamp.and03.ui.core.navigation
+package com.boostcamp.and03.ui.navigation
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
