+++ app/src/main/java/com/boostcamp/and03/ui/screen/addbook/AddBookNavGraph.kt
@@ -3,7 +3,7 @@ package com.boostcamp.and03.ui.screen.addbook
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.boostcamp.and03.ui.core.navigation.MainTabRoute
+import com.boostcamp.and03.ui.navigation.MainTabRoute
 
 fun NavGraphBuilder.addBookNavGraph(
     modifier: Modifier = Modifier
