+++ app/src/main/java/com/boostcamp/and03/ui/screen/mypage/MyPageNavGraph.kt
@@ -3,7 +3,7 @@ package com.boostcamp.and03.ui.screen.mypage
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.boostcamp.and03.ui.core.navigation.MainTabRoute
+import com.boostcamp.and03.ui.navigation.MainTabRoute
 
 fun NavGraphBuilder.myPageNavGraph(
     modifier: Modifier = Modifier
